Fixed GUI Bugs and Reworked Job Submission for Jlag2

+++ model_versions.py
@@ -6,7 +6,8 @@ from command import cmd
 from write_env import write_env
 import jetlag_conf
 import os, sys
-        
+import input_params
+
 class HiddenPrint:
     def __enter__(self):
         self._original_stdout = sys.stdout
@@ -47,35 +48,36 @@ def get_versions(model):
 
     return VERSIONS
 
-def getModelsAndPacks():
+def getModelsAndPacks(update_flag):
     
-    uv = jetlag_conf.get_uv()
-    os.system("rm -fr input.tgz run_dir")
-    os.system("mkdir -p run_dir")
-    print("Retrieving List of Models...")
-    print("This may take a few minutes...")
-    if True: #with HiddenPrint():
-        write_env()
-        os.system("tar czvf input.tgz run_dir")
-        jobid = uv.run_job("get_models_and_packs", nx=1, ny=1, nz=1, jtype="queue", run_time="1:00:00",script_name="getModelsPacks")
-        jobid.wait()
-        uv.get_file(jobid, "run_dir/machineFiles.tar.gz",as_file="machineFiles.tar.gz")
-                
-   
-    #machinePath = "%s/agave-model/machineFiles" % os.environ["HOME"]
-    
-    os.system("rm -rf %s/agave-model/machineFiles" % os.environ["HOME"])
-    os.system("tar -xvf %s/agave-model/machineFiles.tar.gz -C %s/agave-model/" % (os.environ["HOME"], os.environ["HOME"]))
-    os.system("rm %s/agave-model/machineFiles.tar.gz" % os.environ["HOME"])   
+    if (input_params.get("last_jid") != jetlag_conf.get_uv().jetlag_id) or update_flag:
+        uv = jetlag_conf.get_uv()
+        os.system("rm -fr input.tgz run_dir")
+        os.system("mkdir -p run_dir")
+        print("Retrieving List of Models...")
+        print("This may take a few minutes...")
+        if True: #with HiddenPrint():
+            write_env()
+            os.system("tar czvf input.tgz run_dir")
+            jobid = uv.run_job("get_models_and_packs", nx=1, ny=1, nz=1, jtype="queue", run_time="1:00:00",script_name="getModelsPacks")
+            jobid.wait()
+            uv.get_file(jobid, "run_dir/machineFiles.tar.gz",as_file="machineFiles.tar.gz")
+
+
+        #machinePath = "%s/agave-model/machineFiles" % os.environ["HOME"]
+
+        os.system("rm -rf %s/agave-model/machineFiles" % os.environ["HOME"])
+        os.system("tar -xvf %s/agave-model/machineFiles.tar.gz -C %s/agave-model/" % (os.environ["HOME"], os.environ["HOME"]))
+        os.system("rm %s/agave-model/machineFiles.tar.gz" % os.environ["HOME"])   
     
     path = []
     path.append(os.environ["HOME"]+"/agave-model/machineFiles/")
-    path.append(os.environ["HOME"]+"/agave-model/JSONFiles/")
+    path.append(os.environ["HOME"]+"/agave-model/science-models/JSONFiles/")
     
     name_list = []
     package_list = []
     
-    if os.path.isfile(os.environ["HOME"]+"/agave-model/machineFiles/*.json") and os.path.isfile(os.environ["HOME"]+"/agave-model/machineFiles/spack-info.txt") or os.path.isdir(os.environ["HOME"]+"/agave-model/JSONFiles"):
+    if os.path.isfile(os.environ["HOME"]+"/agave-model/machineFiles/*.json") and os.path.isfile(os.environ["HOME"]+"/agave-model/machineFiles/spack-info.txt") or os.path.isdir(os.environ["HOME"]+"/agave-model/science-models/JSONFiles"):
         
         for x in range(2):
             for filename in glob.glob(os.path.join(path[x], '*.json')):
@@ -84,7 +86,8 @@ def getModelsAndPacks():
                     name_list.append(data['name'])
                     package_list.append(data['package'] + '@' + data['version'])
 
-        print("Done!")
+        if input_params.get("last_jid") != jetlag_conf.get_uv().jetlag_id:
+            print("Done!")
         
     else:
         print("Error with retrieving certain files from server!")
